Fix 5 bugs: keyword table overflow, interface class IDs, 16-bit switch, string switch, test gaps
- Lexer: keyword table 40→48 (41 keywords after enum; was OOB on picoJVM)
- Linker: include interface classes as stubs in class table (fixes class ID
  mismatch — CP entries used compiler indices but pjvm table skipped interfaces)
- Compiler: caseVals short[]→int[] (switch on values >32767 was truncated)
- Stmt: implement string switch via equals-dispatch chain (shift-and-insert,
  same pattern as integer LOOKUPSWITCH); detect string type in pSwitch
- Makefile: add T23_StringSwitch and T26_Interface to test suite
- 46/46 tests pass, 45/45 disk tests pass, selfhost fixpoint 33,815B
- Source: 104,754B

+++ picojc/src/Compiler.java
@@ -146,7 +146,7 @@ class C {
 	static int lpDepth;
 
 	// --- Switch case arrays (reused, not nested) ---
-	static short[] caseVals   = new short[64];
+	static int[] caseVals     = new int[64];
 	static short[] caseLbls = new short[64];
 
 	// --- Current context ---
